fix: refactored input managor to not tranform
