+++ main.h
@@ -6,7 +6,6 @@
 #include <Windows.h>
 #include <cstdlib>
 #include <ctime>
-#include <cmath>
 
 // window
 void initWindow(int argc, char** argv);
@@ -15,11 +14,6 @@ void keyboard(unsigned char key, int x, int y);
 void specialKeyboard(int key, int x, int y);
 void mouse(int button, int state, int x, int y);
 void mouseMotion(int x, int y);
-
-// Camera Controll
-void transformCamera();
-void printCameraPosition(int value);
-
 // render
 void display();
 int main(int argc, char** argv);
